Fix cover loading not awaited
This results in the cover image not being sent if it wasn't previously loaded, which might occur when starting a beatmap programatically. This has the potential to send some gameplay related events before the songStart event, but due to a really low chance of that happening, I will not try to mitigate that issue. If you encounter that behavior, feel free to blame Moon for having this issue in the first place and to reach out so we can look into fixing it then.